+++ src/plugins/canvas/mod.rs
@@ -1,2 +1,3 @@
 pub mod client;
 pub mod config;
+pub mod types;
