+++ src/server/mod.rs
@@ -1,3 +1,2 @@
 pub mod error;
 pub mod mcp_server;
-pub mod types;
